Added a mock stripe system to serve as a bounce-point for payment requests. Set up basic DI for the stripe payment classes.

+++ GenericPaymentModule/Models/Payments/Stripe/StripePaymentRequest.cs
@@ -10,7 +10,7 @@ namespace GenericPaymentModule.Models
     /// </summary>
     public class StripePaymentRequest 
     {
-        public decimal Amount;
+        public int Amount;
         public string Currency;
         public decimal ApplicationFee;
         public bool Capture;
